@@ -7,8 +7,11 @@
   },
   "scripts": {
     "start": "tsx server/index.ts",
-    "dev": "tsx watch server/index.ts",
-    "test": "NODE_ENV=test tsx server/test-server.ts"
+    "build": "npm run build:ts",
+    "build:ts": "tsc -p tsconfig.json",
+    "build:clean": "rm -rf dist && npm run build",
+    "validate": "node ai-validation-suite.cjs",
+    "test:all": "node comprehensive-system-test.ts"
   },
   "dependencies": {
     "@anthropic-ai/sdk": "^0.67.0",
@@ -26,7 +29,7 @@
     "cors": "^2.8.5",
     "dotenv": "^16.3.1",
     "drizzle-orm": "^0.29.0",
-    "express": "^4.18.2",
+    "express": "^4.21.2",
     "express-rate-limit": "^7.1.5",
     "express-session": "^1.18.2",
     "helmet": "^7.1.0",
@@ -43,19 +46,9 @@
     "rate-limiter-flexible": "^8.1.0",
     "sharp": "^0.34.4",
     "socket.io": "^4.8.1",
+    "ts-node": "^10.9.2",
     "tsx": "^4.6.0",
-    "typescript": "^5.2.2",
+    "typescript": "^5.9.3",
     "zod": "^3.22.4"
-  },
-  "devDependencies": {
-    "@types/bcryptjs": "^3.0.0",
-    "@types/better-sqlite3": "^7.6.8",
-    "@types/compression": "^1.7.5",
-    "@types/cors": "^2.8.17",
-    "@types/express": "^4.17.21",
-    "@types/jsonwebtoken": "^9.0.5",
-    "@types/node": "^20.9.0",
-    "@types/pdfkit": "^0.13.3",
-    "drizzle-kit": "^0.20.6"
   }
 }
